feat: update chirp methods to use auth
Fix:
- user and authorize functions are undefined

+++ app/Http/Controllers/ChirpController.php
@@ -4,12 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Models\Chirp;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ChirpController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index()
     {
         $chirps = Chirp::with('user')
@@ -20,58 +18,45 @@ class ChirpController extends Controller
         return view('home', ['chirps' => $chirps]);
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
         //
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
+    // Store chirp
     public function store(Request $request)
     {
         // Validate the request
         $validated = $request->validate([
             'message' => 'required|string|max:255',
-        ], [
-            'message.required' => 'Please write something to chirp!',
-            'message.max' => 'Chirps must be 255 characters or less',
         ]);
 
-        // Create a chirp (no user)
-        Chirp::create([
-            'message' => $validated['message'],
-            'user_id' => null,
-        ]);
+        // Use auth user
+        auth()->user()->chirps()->create($validated);
 
         return redirect('/')->with('success', 'Your chirp has been posted!');
     }
 
-    /**
-     * Display the specified resource.
-     */
+    // Display chirp
     public function show(string $id)
     {
         //
     }
 
-    /**
-     * Show the form for editing the specified resource.
-     */
+    // Edit chirp
     public function edit(Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         // Add authorization
         return view ('chirps.edit', compact('chirp'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     */
+    /// Update chirp
     public function update(Request $request, Chirp $chirp)
     {
+        $this->authorize('update', $chirp);
+
         // Validate
         $validated = $request->validate([
             'message' => 'required|string|max:255',
@@ -83,11 +68,11 @@ class ChirpController extends Controller
         return redirect('/')->with('success', 'Chirp updated!');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
+    // Remove chirp
     public function destroy(Chirp $chirp)
     {
+        $this->authorize('delete', $chirp);
+
         $chirp->delete();
 
         return redirect('/')->with('success', 'Chirp deleted!');
